Improve tracking precision - experiment
The tracker used batching and upated every couple of minutes, but using
it has shown that it very often loses precision (location updates with
presicion ~ 30 km are useless for what we want to do here) or stops
working altogether as once it loses its GPS fix it can't get it again
in remote places.

Experiments have shown that it works much better with continuous updates
(like for navigation) and periodic updates done by custom-made time
measurements. This uses more battery, but not as much more as one would
expect, so it seems to be well worth it - locations are almost always
available and very precise most of the time (~ 4-10 m) and update
frequency is stable and independent of other applications. The GPS fix
seems never to be lost either.

This makes the code a bit easier as it may assume every update will just
contain a single location.